+++ osrframework/utils/config_credentials.py
@@ -1,9 +1,11 @@
 # !/usr/bin/python
 # -*- coding: cp1252 -*-
 #
-################################################################################
+##################################################################################
 #
-#    This program is part of OSRFramework. You can redistribute it and/or modify
+#    Copyright 2016 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
+#
+#    This file is part of OSRFramework. You can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
@@ -16,11 +18,11 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-################################################################################
+##################################################################################
 
 import ConfigParser
 import os
-import osrframework.utils.general as general
+import osrframework.utils.configuration as configuration
 
 def returnListOfCreds():
     '''
@@ -32,13 +34,13 @@ def returnListOfCreds():
     '''
     listCreds = []
     # If a accounts.cfg has not been found, creating it by copying from default
-    configPath = general.getConfigPath("accounts.cfg")
+    configPath = configuration.getConfigPath("accounts.cfg")
 
     # Checking if the configuration file exists
     if not os.path.exists(configPath):
         try:
             # Copy the data from the default folder
-            defaultConfigPath = general.getConfigPath(os.path.join("default", "accounts.cfg"))
+            defaultConfigPath = configuration.getConfigPath(os.path.join("default", "accounts.cfg"))
      
             with open(configPath, "w") as oF:
                 with open(defaultConfigPath) as iF: